eatzoneapp/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
Deprecated==1.2.13
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.15.0


Client id: 5NT0uMw6VYzstC9kpjXp0mxAmgQDlSJEoVlKJ5O0
Client secret: WrT2KclVokW8Rp7q9qfiTzmBNHLMUjVynJ6XchDhdhveuLid2TjnxC0Or851kCqC0SIhvQnFM2Qx1BZNUCnsHFzIswx48AU0vZ6lK1Vb2bKow30Fo0pW1p7P5bmLc7TJ
